fix: Entity와 Domain 동기화
복합 키를 위해 존재하는 클래스와 레코드가 복합키가 사용되는 Entity와 자료형이 일치하지 않는 문제를 인지하여 자료형을 수정하는 조치를 단행하였습니다

+++ src/main/java/com/ampersand/groom/domain/booking/domain/constant/TimeSlotId.java
@@ -1,4 +1,4 @@
 package com.ampersand.groom.domain.booking.domain.constant;
 
-public record TimeSlotId(Integer placeId, String timeLabel) {
+public record TimeSlotId(Long placeId, String timeLabel) {
 }
